+++ tests/Fields/TaxonomyTest.php
@@ -36,28 +36,40 @@ class TaxonomyTest extends TestCase
         Taxonomy::make('Invalid Taxonomy Appearance')->appearance('test')->toArray();
     }
 
-    public function testCreateTerms()
+    public function testCanAddTerm()
     {
-        $field = Taxonomy::make('Taxonomy Create Terms')->createTerms()->toArray();
+        $field = Taxonomy::make('Taxonomy Can Add Term')->addTerm(true)->toArray();
         $this->assertTrue($field['add_term']);
     }
 
-    public function testDisableCreateTerms()
+    public function testCannotAddTerm()
     {
-        $field = Taxonomy::make('Taxonomy Undefined Create Terms')->createTerms(false)->toArray();
+        $field = Taxonomy::make('Taxonomy Cannot Add Term')->addTerm(false)->toArray();
         $this->assertFalse($field['add_term']);
     }
 
-    public function testLoadTerms()
+    public function testShouldLoadTerms()
     {
-        $field = Taxonomy::make('Taxonomy Load Terms')->loadTerms()->toArray();
+        $field = Taxonomy::make('Taxonomy Should Load Terms')->loadTerms(true)->toArray();
         $this->assertTrue($field['load_terms']);
     }
 
-    public function testSaveTerms()
+    public function testShouldNotLoadTerms()
     {
-        $field = Taxonomy::make('Taxonomy Save Terms')->saveTerms()->toArray();
+        $field = Taxonomy::make('Taxonomy Should Not Load Terms')->loadTerms(false)->toArray();
+        $this->assertFalse($field['load_terms']);
+    }
+
+    public function testShouldSaveTerms()
+    {
+        $field = Taxonomy::make('Taxonomy Should Save Terms')->saveTerms(true)->toArray();
         $this->assertTrue($field['save_terms']);
+    }
+
+    public function testShouldNotSaveTerms()
+    {
+        $field = Taxonomy::make('Taxonomy Should Not Save Terms')->saveTerms(false)->toArray();
+        $this->assertFalse($field['save_terms']);
     }
 
     public function testTaxonomy()